requirements.txt
librosa==0.10.1
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
soundfile==0.12.1
streamlit==1.28.0
faiss-cpu==1.7.4